refactor: remove upward dependencies from the ui layer
