+++ src/components/CategorizationRules/CategorizationRulesTable/CategorizationRulesTable.tsx
@@ -3,13 +3,14 @@ import type { Row } from '@tanstack/react-table'
 import { Pencil, Trash2 } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
+import { type TablePaginationProps } from '@internal-types/utility/pagination'
 import type { CategorizationRule } from '@schemas/bankTransactions/categorizationRules/categorizationRule'
 import type { NestedCategorization } from '@schemas/categorization'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { Button } from '@ui/Button/Button'
 import { Span } from '@ui/Typography/Text'
 import type { ColumnConfig } from '@blocks/DataTable/utils/column'
-import { PaginatedTable, type TablePaginationProps } from '@blocks/PaginatedDataTable/PaginatedDataTable'
+import { PaginatedTable } from '@blocks/PaginatedDataTable/PaginatedDataTable'
 import { ResolvedCategoryName } from '@components/CategorizationRules/ResolvedCategoryName'
 import { getCategorizationRuleAmountLabel, getCategorizationRuleCounterpartyLabel, getCategorizationRuleDirectionLabel } from '@components/CategorizationRules/utils'
