+++ src/components/features/invoices/InvoiceTable/InvoiceTable.tsx
@@ -4,6 +4,7 @@ import type { TFunction } from 'i18next'
 import { Plus } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
+import { type TablePaginationProps } from '@internal-types/utility/pagination'
 import { type Invoice, InvoiceStatus } from '@schemas/invoices/invoice'
 import { getCustomerName } from '@utils/customer'
 import { unsafeAssertUnreachable } from '@utils/switch/assertUnreachable'
@@ -19,7 +20,6 @@ import { DataTableHeader } from '@blocks/DataTable/DataTableHeader'
 import type { ColumnConfig } from '@blocks/DataTable/utils/column'
 import { Container } from '@blocks/layout/Container/Container'
 import { PaginatedTable } from '@blocks/PaginatedDataTable/PaginatedDataTable'
-import { type TablePaginationProps } from '@blocks/PaginatedDataTable/PaginatedDataTable'
 import { InvoiceStatusCell } from '@features/invoices/InvoiceStatusCell/InvoiceStatusCell'
 import { useInvoiceStatusOptions } from '@features/invoices/utils'
