+++ src/components/features/mileage/TripsTable/TripsTable.tsx
@@ -4,6 +4,7 @@ import type { TFunction } from 'i18next'
 import { Edit, Trash2 } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
+import { type TablePaginationProps } from '@internal-types/utility/pagination'
 import { type Trip } from '@schemas/trip'
 import { formatCalendarDate } from '@utils/time/timeUtils'
 import { getVehicleDisplayName } from '@utils/vehicles'
@@ -13,7 +14,7 @@ import { HStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
 import type { ColumnConfig } from '@blocks/DataTable/utils/column'
 import { Container } from '@blocks/layout/Container/Container'
-import { PaginatedTable, type TablePaginationProps } from '@blocks/PaginatedDataTable/PaginatedDataTable'
+import { PaginatedTable } from '@blocks/PaginatedDataTable/PaginatedDataTable'
 import { TripsAddressCell } from '@features/mileage/TripAddressCell/TripAddressCell'
 import { TripsTableHeader } from '@features/mileage/TripsTableHeader/TripsTableHeader'
 import { formatDistance, getPurposeLabel } from '@features/mileage/utils'
