+++ src/components/features/timeTracking/TimeEntriesTable/TimeEntriesTable.tsx
@@ -4,6 +4,7 @@ import type { TFunction } from 'i18next'
 import { Edit, Trash2 } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
+import { type TablePaginationProps } from '@internal-types/utility/pagination'
 import type { TimeEntry } from '@schemas/timeTracking'
 import { formatCalendarDate } from '@utils/time/timeUtils'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
@@ -14,7 +15,7 @@ import { DurationSpan } from '@ui/Typography/DurationSpan'
 import { Span } from '@ui/Typography/Text'
 import type { ColumnConfig } from '@blocks/DataTable/utils/column'
 import { Container } from '@blocks/layout/Container/Container'
-import { PaginatedTable, type TablePaginationProps } from '@blocks/PaginatedDataTable/PaginatedDataTable'
+import { PaginatedTable } from '@blocks/PaginatedDataTable/PaginatedDataTable'
 import { TimeEntriesTableHeader } from '@features/timeTracking/TimeEntriesTableHeader/TimeEntriesTableHeader'
 
 import './timeEntriesTable.scss'
