+++ src/hooks/legacy/useJournal.tsx
@@ -1,11 +1,11 @@
 import { useCallback, useEffect, useState } from 'react'
 
 import { SortOrder } from '@internal-types/utility/pagination'
+import { type TablePaginationProps } from '@internal-types/utility/pagination'
 import { type LedgerEntry } from '@schemas/generalLedger/ledgerEntry'
 import { LedgerEntriesSortBy, type ListLedgerEntriesReturn, useGetListLedgerEntries } from '@api/businesses/[business-id]/ledger/entries/get'
 import { useTablePaginationProps } from '@hooks/utils/pagination/useTablePaginationProps'
 import { useLedgerDateRange } from '@providers/DateStoreProvider/LedgerDateStoreProvider'
-import { type TablePaginationProps } from '@blocks/PaginatedDataTable/PaginatedDataTable'
 
 export const JOURNAL_PAGE_SIZE = 15
