+++ src/schemas/common/table.ts
@@ -1,12 +1,7 @@
 import { Schema } from 'effect'
 
+import { Alignment, Pinning } from '@internal-types/utility/table'
 import { createTransformedEnumSchema } from '@schemas/utils'
-
-export enum Alignment {
-  Left = 'LEFT',
-  Right = 'RIGHT',
-  Center = 'CENTER',
-}
 
 const AlignmentSchema = Schema.Enums(Alignment)
 
@@ -15,12 +10,6 @@ export const TransformedAlignmentSchema = createTransformedEnumSchema(
   Alignment,
   Alignment.Left,
 )
-
-export enum Pinning {
-  Left = 'LEFT',
-  Right = 'RIGHT',
-  Unpinned = 'UNPINNED',
-}
 
 const PinningSchema = Schema.Enums(Pinning)
