+++ src/types/toast.ts
@@ -1,0 +1,7 @@
+export interface ToastProps {
+  id?: string
+  content: string
+  duration?: number
+  isExiting?: boolean
+  type?: 'success' | 'error' | 'default'
+}
