+++ src/types/utility/table.ts
@@ -1,0 +1,11 @@
+export enum Alignment {
+  Left = 'LEFT',
+  Right = 'RIGHT',
+  Center = 'CENTER',
+}
+
+export enum Pinning {
+  Left = 'LEFT',
+  Right = 'RIGHT',
+  Unpinned = 'UNPINNED',
+}
